Split CLI smoke and contract tests
